chore: change result to an entity
